+++ Maple2.File.Parser/Xml/Item/Skill.cs
@@ -3,6 +3,6 @@
 namespace Maple2.File.Parser.Xml.Item {
     public class Skill {
         [XmlAttribute] public int skillID;
-        [XmlAttribute] public int skillLevel = 1;
+        [XmlAttribute] public short skillLevel = 1;
     }
 }
